+++ lib/Windows/window_printer_manager.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:developer';
 import 'dart:typed_data';
 
 import 'package:ffi/ffi.dart';
@@ -47,42 +46,14 @@ class WindowPrinterManager {
 
   StreamSubscription? subscription;
 
-  // Find all BLE devices
-  Future<void> startscan() async {
-    if (!isInitialized) {
-      log("Init");
-      await init();
-    }
-    if (!isInitialized) {
-      throw Exception(
-        'WindowBluetoothManager is not initialized. Try starting the scan again',
-      );
-    }
-    List<Printer> devices = [];
-    WinBle.startScanning();
-    subscription = WinBle.scanStream.listen((device) async {
-      log(device.name);
-      devices.add(Printer(
-        address: device.address,
-        name: device.name,
-        connectionType: ConnectionType.BLE,
-        isConnected: await WinBle.isPaired(device.address),
-        // isConnected: false,
-      ));
-    });
-  }
-
   // Connect to a BLE device
   Future<bool> connect(Printer device) async {
     if (!isInitialized) {
       throw Exception('WindowBluetoothManager is not initialized');
     }
-    bool isConnected = false;
-    final subscription = WinBle.connectionStream.listen((device) {});
     await WinBle.connect(device.address!);
-    await Future.delayed(const Duration(seconds: 3));
-    subscription.cancel();
-    return isConnected;
+    await Future.delayed(const Duration(seconds: 5));
+    return await WinBle.isPaired(device.address!);
   }
 
   // Print data to a BLE device
@@ -157,7 +128,28 @@ class WindowPrinterManager {
     ],
   }) async {
     List<Printer> btlist = [];
-    if (connectionTypes.contains(ConnectionType.BLE)) {}
+    if (connectionTypes.contains(ConnectionType.BLE)) {
+      await init();
+      if (!isInitialized) {
+        await init();
+      }
+      if (!isInitialized) {
+        throw Exception(
+          'WindowBluetoothManager is not initialized. Try starting the scan again',
+        );
+      }
+      WinBle.stopScanning();
+      WinBle.startScanning();
+      subscription?.cancel();
+      subscription = WinBle.scanStream.listen((device) async {
+        btlist.add(Printer(
+          address: device.address,
+          name: device.name,
+          connectionType: ConnectionType.BLE,
+          isConnected: await WinBle.isPaired(device.address),
+        ));
+      });
+    }
     List<Printer> list = [];
     if (connectionTypes.contains(ConnectionType.USB)) {
       _usbSubscription?.cancel();